Crate rsp2_sparse

Basic definitions of sparse matrix formats and conversions between them.

These are currently pretty loose abstractions with public data members. There is not a lot of code in rsp2 that requires sparse matrices, and seldom is the same operation needed more than once.

NOTE: These types might even be used with T = M33 or similar to represent a block format, which is totally cool for a storage format but could easily lead to generic mathematical operations doing the wrong thing (if written with poorly chosen generic bounds). That’s why they don’t provide very many mathematical operations.

Structs

RawBee
BTreeMap-based representation.

RawCoo
Coordinate format.

RawCsr
Compressed sparse row.

SparseMatrixError
Returned by validate to indicate that a sparse matrix is invalid.