Struct rsp2_sparse::RawCoo

pub struct RawCoo<T, R: Idx = usize, C: Idx = R> {
    pub dim: (usize, usize),
    pub val: Vec<T>,
    pub row: Vec<R>,
    pub col: Vec<C>,
}

Coordinate format.

Elements in arbitrary order. Duplicates are allowed; they are implicitly summed. Great for building a matrix as a sum of terms, but otherwise impossible to work with.

This type is “Raw” because its members are public and its invariants are not protected. It exists to document the “intent” of some fields of a struct, and to provide some simple utility methods.

Fields
dim: (usize, usize)val: Vec<T>row: Vec<R>col: Vec<C>
Implementations

impl<T, R: Idx, C: Idx> RawCoo<T, R, C>

pub fn validate(&self) -> Result<(), SparseMatrixError>

pub fn map<F, T2>(self, f: F) -> RawCoo<T2, R, C> where
F: FnMut(T) -> T2,

pub fn map_row_indices<F, R2: Idx>(
self,
new_dim: usize,
f: F
) -> RawCoo<T, R2, C> where
F: FnMut(R) -> R2,

COO provides the unique capability of mapping indices with an arbitrary function. Remember, entries with duplicate indices in the output will be implicitly summed.

pub fn map_column_indices<F, C2: Idx>(
self,
new_dim: usize,
f: F
) -> RawCoo<T, R, C2> where
F: FnMut(C) -> C2,

pub fn into_raw_transpose(self) -> RawCoo<T, C, R>

Transpose R and C, disregarding T. (i.e. if T happens to be some kind of matrix block type like M33, it will not be affected by this operation)

pub fn nnz_with_duplicates(&self) -> usize

impl<T, R: Idx, C: Idx> RawCoo<T, R, C>

pub fn into_csr(self) -> RawCsr<T, R, C> where
T: AddAssign,

pub fn into_bee(self) -> RawBee<T, R, C> where
T: AddAssign,

pub fn into_indexed_dense(self) -> Indexed<R, Vec<Indexed<C, Vec<T>>>> where
T: Zero + AddAssign,

pub fn into_dense(self) -> Vec<Vec<T>> where
T: Zero + AddAssign,

pub fn into_csr_with<F>(self, add_assign: F) -> RawCsr<T, R, C> where
F: FnMut(&mut T, T),

pub fn into_bee_with<F>(self, add_assign: F) -> RawBee<T, R, C> where
F: FnMut(&mut T, T),

pub fn into_indexed_dense_with<Z, F>(
self,
zero: Z,
add_assign: F
) -> Indexed<R, Vec<Indexed<C, Vec<T>>>> where
Z: FnMut() -> T,
F: FnMut(&mut T, T),

pub fn into_dense_with<Z, F>(self, zero: Z, add_assign: F) -> Vec<Vec<T>> where
Z: FnMut() -> T,
F: FnMut(&mut T, T),

pub fn from_dense(mat: Vec<Vec<T>>) -> Self where
T: Zero,

impl<T, R: Idx, C: Idx> RawCoo<T, R, C> where
T: Clone,

pub fn to_raw_transpose(&self) -> RawCoo<T, C, R>

pub fn to_csr(&self) -> RawCsr<T, R, C> where
T: AddAssign,

pub fn to_bee(&self) -> RawBee<T, R, C> where
T: AddAssign,

pub fn to_indexed_dense(&self) -> Indexed<R, Vec<Indexed<C, Vec<T>>>> where
T: Zero + AddAssign,

pub fn to_dense(&self) -> Vec<Vec<T>> where
T: Zero + AddAssign,

pub fn to_csr_with<F>(&self, add_assign: F) -> RawCsr<T, R, C> where
F: FnMut(&mut T, T),

pub fn to_bee_with<F>(&self, add_assign: F) -> RawBee<T, R, C> where
F: FnMut(&mut T, T),

pub fn to_indexed_dense_with<Z, F>(
&self,
zero: Z,
add_assign: F
) -> Indexed<R, Vec<Indexed<C, Vec<T>>>> where
Z: FnMut() -> T,
F: FnMut(&mut T, T),

pub fn to_dense_with<Z, F>(&self, zero: Z, add_assign: F) -> Vec<Vec<T>> where
Z: FnMut() -> T,
F: FnMut(&mut T, T),